requirements.txt
asgiref==3.8.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-log==0.4.0
cloudinary==1.41.0
cloudinary-cli==1.10.3
Django==5.1.1
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
docstring_parser==0.16
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.4
MarkupSafe==2.1.5
maskpass==0.3.7
oauthlib==3.2.2
pillow==10.4.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.18.0
pynput==1.7.6
pyobjc-core==10.0
pyobjc-framework-ApplicationServices==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.0.7
